Routes now handled by middleware
