+++ server/routes/login.js
@@ -1,0 +1,8 @@
+var express = require("express");
+var router = express.Router();
+
+router.get("/", function(request, response) {
+  response.render("login");
+});
+
+module.exports = router;
