+++ server/views/login.handlebars
@@ -1,0 +1,72 @@
+<html>
+
+<head>
+  <meta charset="utf-8" />
+  <title>Food Waste Prevention Login</title>
+  <style>
+    .login-form {
+      width: 300px;
+      margin: 0 auto;
+      font-family: Tahoma, Geneva, sans-serif;
+    }
+
+    .login-form h1 {
+      text-align: center;
+      color: #4d4d4d;
+      font-size: 24px;
+      padding: 20px 0 20px 0;
+    }
+
+    .login-form input[type="password"],
+    .login-form input[type="text"] {
+      width: 100%;
+      padding: 15px;
+      border: 1px solid #dddddd;
+      margin-bottom: 15px;
+      box-sizing: border-box;
+    }
+
+    .login-form input[type="submit"] {
+      width: 100%;
+      padding: 15px;
+      background-color: #535b63;
+      border: 0;
+      box-sizing: border-box;
+      cursor: pointer;
+      font-weight: bold;
+      color: #ffffff;
+    }
+
+    .account-button {
+      width: 25%;
+      position: relative;
+      left: 37.5%;
+      padding: 10px;
+      background-color: #24303b;
+      border: 0;
+      box-sizing: border-box;
+      cursor: pointer;
+      font-weight: bold;
+      color: #ffffff;
+    }
+  </style>
+</head>
+
+<body>
+  <div class="login-form">
+    <h1>Food Waste Prevention</h1>
+    <form action="auth" method="POST">
+      <input type="text" name="username" placeholder="Username" required />
+      <input type="password" name="password" placeholder="Password" required />
+      <input type="submit" />
+    </form>
+  </div>
+  <br />
+  <div>
+    <form action="create-account" method="GET">
+      <input class="account-button" type="submit" value="Create Account" />
+    </form>
+  </div>
+</body>
+
+</html>
